simulator/requirements.txt
-i https://pypi.org/simple
confluent-kafka==1.7.0
numpy==2.3.2; python_version >= '3.12'
pandas==2.3.2; python_version >= '3.9'
pyarrow==21.0.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.2
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tk==0.1.0
tzdata==2025.2; python_version >= '2'
